Update analysis.py
Made useful comments throughout entirety of code. Included the addition of command line arguments to change kNN and the number of generated signal and bkg events if desired. Also defined functions to visualize the signal and bkg fits produced for an event and its nearest neighbors. Lastly, included parallelization of the generation of single and bkg events if desired to speed up processes.