Update analysis.py
Updates the code to use local density when calculating knn as well as use a fixed radius to determine a neighborhood of events. Each method handles the issue of leaving k fixed and allows for a variation depending on density of surrounding events nearby. These methods seem to correct issues regarding keeping k fixed since different aspects of phase-space are more or less dense than others, meaning k should vary depending on the density of events in a specific region of phase-space